Multiple changes to shiphawk gem
Updated main class to allow api_token to be set through the shiphawk config block.
Updated main class to allow Shiphawk to be used as a singleton
Added experimentary update_catalog_data method.
Updated the shipments_status_update method to not take an id. (you pass it a list of ids)